`origin/runtime-values-cuda` (794ff86b) is `origin/main` (da6ba18c) plus the
Dell's six commits, so this single merge brings both: the 167 trunk commits
metal-first-light was behind, and the Dell's cherry-picks of my runtime-values
work with its corrections on top.

Four conflicts, each resolved deliberately:

* `src/neurobrix/core/prism/solver.py` -> THE DELL'S, wholesale. It owns
  core/prism and its version is the corrected one. Two real corrections to my
  `d459403c`: a declared symbol is now put into the SYMBOL's unit before
  dividing (Sana-1600M declares height/width in latent rows with trace_value 32
  against a 1024 trace_resolution, so the raw ratio multiplied the estimate by
  32 per axis and Prism refused a model that had rendered that morning), and one
  DIMENSION now contributes once however many times it is declared
  (PixArt-XL-1024 declares seq_len at two symbol ids and read x6.25 for x2.5).
  It also restores a call site the cherry-pick had orphaned.

* `tests/unit/kernels/test_bench_refuses_to_measure_the_swap.py` -> the Dell's
  framing for the duplicated part, my two additional cases kept on top. We fixed
  the same bug independently within hours; its version straddles
  `available - flush` at 1000 MB and landed first, so it is the duplicate that
  stands. Kept from mine, because they are additions and not duplicates: the
  case where available memory is BELOW the flush (every sweep swaps then,
  arguments or none), and the pin that reads `do_bench`'s source so the gate and
  the allocation it counts cannot drift apart again.

* `tools/apple_matrix_percell.py` -> MINE, and `tools/measured_peaks.json`
  REMOVED. This is the one place where "favour the Dell's version" and the
  measurement disagree, and the measurement wins: the Dell's copy is the
  PRE-correction one. It reads a per-model `measured_peaks.json` table, which
  Hocine ordered removed on 2026-09-17 as model-specific knowledge inside a gate
  ("remove the table"). The cherry-pick carried the older state forward. Apple
  tooling is this machine's to own; the table does not come back.

* `docs/reference/owed-proofs.md` -> UNION. An append-only ledger; neither side
  supersedes the other, and both sets of entries are kept whole.

Audited in both directions rather than assumed. Still present after the merge:
the trailing-buffer seam on all four drivers, `do_bench` allocating at
`_BENCH_FLUSH_BYTES`, the ext driver asking `TG_BUDGET_BYTES`, `MetalDriver`'s
device query, the Metal seam's fallback markers, `runtime_values`. And from the
Dell: `_SYMBOL_UNIT`, `seen_dims`, its bench-door straddle.

`metal_backend: triton_msl` still stands at apple_m4_pro.yml:370 — that is the
next item, not this one.

+++ src/neurobrix/kernels/ops/decode_attn_vec.py
@@ -94,7 +94,8 @@ def decode_attn_vec_split_kernel(
     q = _round_q(q, Q_TO, Q_SATURATE)
 
     seg_start = pid_s * seg_len
-    seg_end = tl.minimum(seg_start + seg_len, T_k)
+    seg_start = seg_start.to(tl.int32)   # loop bounds are 32-bit counts: the Metal lowering refuses a 64-bit scf.for bound (addressing stays 64-bit through the program ids)
+    seg_end = tl.minimum(seg_start + seg_len, T_k).to(tl.int32)
 
     m_i = float("-inf")
     l_i = 0.0
@@ -179,7 +180,8 @@ def decode_attn_vec_grouped_kernel(
     offs_g = tl.arange(0, GQA_GROUPS)
 
     seg_start = pid_s * seg_len
-    seg_end = tl.minimum(seg_start + seg_len, T_k)
+    seg_start = seg_start.to(tl.int32)   # loop bounds are 32-bit counts: the Metal lowering refuses a 64-bit scf.for bound (addressing stays 64-bit through the program ids)
+    seg_end = tl.minimum(seg_start + seg_len, T_k).to(tl.int32)
 
     m_i = tl.full((GQA_GROUPS,), float("-inf"), dtype=tl.float32)
     l_i = tl.zeros((GQA_GROUPS,), dtype=tl.float32)
